Freeze and thaw.
->freeze freezes the current cache into a plain Perl data structure.
This can then be written out however and restored into a new object
later to continue the same sync session.

This doesn't currently freeze the search parameters, the server or the
authentication, all of which are potentially part of the sync session. I
think the server and the auth I will leave up to the client code (I
hardly want to reauth or reconnect, so all I could do is warn if it
appears to be wrong), but I do want to include the search parameters.