Handle Sync Info properly.
This means tracking which state we are in, and properly accumulating
Present notifications and acting on them at the end of the refresh
stage.

I don't yet handle Sync Done properly; in particular, I need logic for
e-syncRefreshRequired, even if I want to normally pass reloadHint to
avoid it. (If I'm caching I don't see any point starting a new search
when I could get the results in this search. If I'm not caching I don't
care about the result, I just pass it back to the user.)