Handle Sync Done controls.
These are only returned at the end of the search; it is important to
handle them correctly in the case of a refreshOnly search where all
entries are marked not Present.

The RFC is unclear about the use of the refreshDeletes member when the
Sync Done is sent at the end of a refreshAndPersist search. The default
is FALSE, but AFAICS this is incorrect if we are in the persist stage:
the current code will call _process_present, even though _present will
not have been populated. I don't believe this will cause a problem if
the server follows the protocol correctly, since there is no way to
leave the refresh stage without clearing _present.